Split mutation_writer back out into it's own module
